Pull out `RatingBlock` from `RatingBlocks` into its own file
This commit puts the `RatingBlock` class into a mixin for the
`RatingBlocks` companion object to extend.  This is done purely to
increase readability by reducing the file size of `RatingBlocks.scala`.
